Set-to-plug: Clear 'controls' entries for new plug key and children
Say you had a plug with args 'x' and 'y'.
Then you set this to a new plug with 'a' and 'b'.
The 'x' and 'y' entries in 'controls' need to be removed.
The 'addControl' logic ensures that the tree will be rebuilt
the next time you try to set a non-plug on one of those branches.